+++ backend/model/historicoModel.go
@@ -4,7 +4,7 @@ import "time"
 
 type Historico struct {
 	ID             uint      `json:"id"              gorm:"primaryKey"`
-	AlunoRM        int       `json:"aluno_rm"        gorm:"not null;column:aluno_rm"`
+	AlunoRM        int64     `json:"aluno_rm"        gorm:"not null;column:aluno_rm"`
 	Tipo           string    `json:"tipo"            gorm:"type:tipo_pagamento;not null"`
 	FormaPagamento string    `json:"forma_pagamento" gorm:"type:forma_pagamento;not null"`
 	Valor          float64   `json:"valor"           gorm:"type:numeric(10,2);not null"`
@@ -16,7 +16,7 @@ func (Historico) TableName() string {
 }
 
 type HistoricoCreate struct {
-	AlunoRM        int     `json:"aluno_rm"        validate:"required"`
+	AlunoRM        int64   `json:"aluno_rm"        validate:"required"`
 	Tipo           string  `json:"tipo"            validate:"required,oneof=credito debito"`
 	FormaPagamento string  `json:"forma_pagamento" validate:"required,oneof=pix boleto saldo"`
 	Valor          float64 `json:"valor"           validate:"required,gt=0"`
@@ -24,7 +24,7 @@ type HistoricoCreate struct {
 
 type HistoricoResponse struct {
 	ID             uint      `json:"id"`
-	AlunoRM        int       `json:"aluno_rm"`
+	AlunoRM        int64     `json:"aluno_rm"`
 	Tipo           string    `json:"tipo"`
 	FormaPagamento string    `json:"forma_pagamento"`
 	Valor          float64   `json:"valor"`
